Add tests for getPasswordEntry function.
We implement a testing class that inherits from Google Test class. The testing
class provides a common setup and teardown mechanism for multiple test cases,
that read from std::cin. Inheritance is public as Google Test requires
SetUp() and TearDown() functions to be public. For oldCin we use a raw
pointer instead of a smart pointer, as the std::streambuf object is owned by
std::cin and smart pointer would try to delete this object, when it goes out
of scope, but we restore the original buffer after the test.